requirements.txt
absl-py==0.13.0
aioice==0.7.5
aiortc==1.2.1
alabaster==0.7.12
anyio==3.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.1.1
astroid==2.6.4
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.5
av==8.0.3
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
binaryornot==0.4.4
black==21.7b0
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
clang==5.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cookiecutter==1.7.3
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
dnspython==2.1.0
docutils==0.17.1
dpcontracts==0.6.0
entrypoints==0.3
flake8==3.9.2
Flask==1.1.4
flatbuffers==1.12
forbiddenfruit==0.1.4
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-crc32c==1.1.2
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
hyperopt==0.2.5
idna==2.10
imagesize==1.2.0
importlib-metadata==4.6.1
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
IProgress==0.4
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.2
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.7.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.5.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==23.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lightgbm==3.2.1
loguru==0.5.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
names==0.3.0
nbclassic==0.2.8
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.11.0
networkx==2.5.1
notebook==6.4.0
numpy==1.19.5
numpydoc==1.1.0
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pathspec==0.9.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
poyo==0.5.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.1.1
pyee==8.1.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==1.7.1
pylibsrtp==0.6.8
pylint==2.9.3
pyls-black==0.4.7
pyls-spyder==0.4.0
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.9.0
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.2
python-slugify==5.0.2
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
QDarkStyle==3.0.2
qstylizer==0.2.0
QtAwesome==1.0.3
qtconsole==5.1.1
QtPy==1.9.0
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rope==0.19.0
rsa==4.7.2
Rtree==0.9.7
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
Sphinx==4.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlitedict==1.7.0
statsmodels==0.12.2
syft==0.5.0
syft-proto==0.5.3
tenseal==0.3.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.1.0
three-merge==0.1.1
tinycss2==1.1.0
toml==0.10.2
tomli==1.0.4
torch==1.8.1
torchcsprng==0.2.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win32-setctime==1.0.3
wrapt==1.12.1
xgboost==1.4.2
yapf==0.31.0
zipp==3.5.0
